+++ src/options/main.tsx
@@ -3,10 +3,14 @@ import '@/index.css'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 
+import { TooltipProvider } from '@/shared/components/ui/tooltip'
+
 import App from './app.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <TooltipProvider>
+      <App />
+    </TooltipProvider>
   </StrictMode>,
 )